WIP vvvv SEE comment below:
Add field-specifiers for resource-changing handlers

The logic to narrow old/new/diff to that field already exists
in the codebase, and it is used for the `on.field` handlers.

Now, it will be usable for resume-/create-/update-/delete-handlers.
These handlers will be selected only if that field is affected,
and ignored otherwise.